nepenthes  - submit-postgres    - compile fix on openbsd  - module-peiros    - fix on every os != linux  - module-honeytrap    - #ifdef HAVE_PCAP    - remove useless (empty) pcap files in PCAPSocket destruktor

+++ modules/module-peiros/PeirosParser.cpp
@@ -238,7 +238,7 @@ string PeirosParser::renderRequest(PeirosRequest * request)
 	{
 		char * lengthHeader;
 		
-		asprintf(&lengthHeader, "Content-length: %u\r\n", request->appendedData.size());
+		asprintf(&lengthHeader, "Content-length: %u\r\n", (int)request->appendedData.size());
 		result += lengthHeader;
 		free(lengthHeader);
 	}
